backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
ffmpeg-python==0.2.0
openai-whisper @ git+https://github.com/openai/whisper.git
ctranslate2==4.3.1
sentencepiece==0.2.0
numpy==1.26.4
torch==2.4.0